Update Conway
This is the first Beta version of the program. It fully implements
the rules of Conway's Game of Life and uses a toroidal array to wrap
around the grid. Full disclosure: I tried my best to come up with an
algorithm that would correctly scan the neighbors of every cell,
including the cells on the edge of the grid. I could not, and 
eventually searched for one and found the one used in this
implementation: http://en.wikibooks.org/wiki/Algorithm_Implementation/Simulation/Conway's_Game_of_Life
That algorithm is the one used in my implementation. I did not
create it.